// lib/api-client/client/types.gen.ts
export interface Config<T extends ClientOptions = ClientOptions>
  extends Omit<RequestInit, "body" | "headers" | "method">, CoreConfig {
  /**
   * Base URL for all requests made by this client.
   */
  baseUrl?: T["baseUrl"];
  /**
   * Fetch API implementation. You can use this option to provide a custom
   * fetch instance.
   *
   * @default globalThis.fetch
   */
  fetch?: typeof fetch;
  /**
   * Return the response data parsed in a specified format. By default, `auto`
   * will infer the appropriate method from the `Content-Type` response header.
   * You can override this behavior with any of the {@link Body} methods.
   * Select `stream` if you don't want to parse response data at all.
   *
   * @default 'auto'
   */
  parseAs?:
    | "arrayBuffer"
    | "auto"
    | "blob"
    | "formData"
    | "json"
    | "stream"
    | "text";
  /**
   * Throw an error instead of returning it in the response?
   *
   * @default false
   */
  throwOnError?: T["throwOnError"];
}

export interface RequestOptions<
  TData = unknown,
  ThrowOnError extends boolean = boolean,
  Url extends string = string,
>
  extends
// ...
export interface ResolvedRequestOptions<
  ThrowOnError extends boolean = boolean,
  Url extends string = string,
> extends RequestOptions<unknown, ThrowOnError, Url> {
  headers: Headers;
  serializedBody?: string;
}

export type RequestResult<
  TData = unknown,
  TError = unknown,
  ThrowOnError extends boolean = boolean,
> = ThrowOnError extends true
  ? Promise<{
      error: any;
      data: TData extends Record<string, unknown> ? TData[keyof TData] : TData;
      response: Response;
    }>
  : Promise<
      (
        | {
            data: TData extends Record<string, unknown>
              ? TData[keyof TData]
              : TData;
            error: undefined;
          }
        | {
            data: undefined;
            error: TError extends Record<string, unknown>
              ? TError[keyof TError]
              : TError;
          }
      ) & {
        /** response may be undefined due to a network error where no response object is produced */
        response?: Response;
      }
    >;

export interface ClientOptions {
  baseUrl?: string;
  throwOnError?: boolean;
}

type MethodFn = <
  TData = unknown,
  TError = unknown,
  ThrowOnError extends boolean = false,
>(
  options: Omit<RequestOptions<TData, ThrowOnError>, "method">,
) => RequestResult<TData, TError, ThrowOnError>;

type SseFn = <
  TData = unknown,
  // eslint-disable-next-line @typescript-eslint/no-unused-vars
  _TError = unknown,
  ThrowOnError extends boolean = false,
>(
  options: Omit<RequestOptions<never, ThrowOnError>, "method">,
) => Promise<ServerSentEventsResult<TData>>;

type RequestFn = <
  TData = unknown,
  TError = unknown,
  ThrowOnError extends boolean = false,
>(
  options: Omit<RequestOptions<TData, ThrowOnError>, "method"> &
    Pick<Required<RequestOptions<TData, ThrowOnError>>, "method">,
) => RequestResult<TData, TError, ThrowOnError>;

type BuildUrlFn = <
  TData extends {
    body?: unknown;
    path?: Record<string, unknown>;
    query?: Record<string, unknown>;
    url: string;
  },
>(
  options: TData & Options<TData>,
) => string;

export type Client = CoreClient<
  RequestFn,
  Config,
  MethodFn,
  BuildUrlFn,
  SseFn
> & {
  interceptors: Middleware<Response, unknown, ResolvedRequestOptions>;
};
// ...
export type Options<
  TData extends TDataShape = TDataShape,
  ThrowOnError extends boolean = boolean,
  TResponse = unknown,
> = OmitKeys<
  RequestOptions<TResponse, ThrowOnError>,
  "body" | "path" | "query" | "url"
> &
  ([TData] extends [never] ? unknown : Omit<TData, "url">);